requirements.txt
aiohttp
aiosqlite
discord.py
python-dotenv
factorio-rcon-py
anyio
